combine tests for create group and user
